lint: the four golangci findings #909's lineage code left, and the five modernize names behind them
golangci-lint, on the lock's lineage file and its test:

  - errorlint: isLineageDrift type-asserted on the error, which misses a
    wrapped lineageDrift. errors.As, which is what the check asks for and what
    the function means;
  - gocritic appendAssign: `lv.Lineage = append(history, cur)` appends to one
    slice and assigns to another. The append now lands on `history` and
    `history` is assigned — the same two statements with the intent said out
    loud;
  - godot, twice: a test's doc comment ended on a closing quote, because the
    spec sentence it quotes ends in a period. The period moves outside the
    quote, as the other seventeen spec-quoting comments in the tree do.

AND THE FIVE MODERNIZE NAMES UNDER THEM, which no run had reached yet: the lint
job runs modernize AFTER golangci-lint, so #909's lockfile findings were hidden
behind the four above and would have turned the job red a second time the moment
it got past them. A `for i := 0; i < 3; i++` becomes `range 3`, hasValue becomes
slices.Contains, and three `range strings.Split` become `range strings.SplitSeq`.

The unconvert finding is this branch's own: the ordinal guard in the commit
before this one converted an int64 to int64.

+++ internal/codegen/cpptable/fixedform.go
@@ -853,7 +853,7 @@ func (g *tableGen) emitFixedClampElement(f *ir.Field, expr string, indent int) {
 			// above the extent, so the comparison is always false and the
 			// emitter drops it rather than hand a warning-as-error build a
 			// tautology. No semantics move — the elided check never clamped.
-			if ordinalFillsStorage(int64(r.Max), r.StorageBits) {
+			if ordinalFillsStorage(r.Max, r.StorageBits) {
 				return
 			}
 			g.pf("%sif ( (uint64_t) %s > %du ) { %s = %s::None; clamped++; }\n",
